requirements.txt
pandas
numpy
typer
random
time
datetime
ccxt
binance-python
yaml
requests
statistics